created modules folder and moved rabbitmq there
